+++ NEWS.md
@@ -14,6 +14,7 @@
 * the `resize` argument in `procimg()` now takes a percentage, rather than scaling factor
 * fixed a bug where `colspace(space = segment)` would return both a `B` and (redundant) `lum` column
 * fixed a bug where the rod sensitivity of _Canis familiaris_ was inaccessible through `vismodel()`
+* fixed an issue in `getspec()` where badly encoded characters in some spectral files would cause failure
 
 # pavo 2.1.0